+++ app/components/output/generic-table/generic-table.component.ts
@@ -128,46 +128,12 @@ export class GenericTableComponent implements AfterViewChecked, OnChanges {
         return this.sanitization.bypassSecurityTrustStyle(column.type.colors.getColor(label));
     }
 
-    public getForegroundColor(color: any): SafeHtml {
-        const h = Number.parseInt(color['changingThisBreaksApplicationSecurity'].substr(4, 3).split(',')[0]);
-        const s = 1;
-        const l = 0.7;
-        const rgb = this.hslToRgb(h, s, l);
-        const yiq = ((rgb[0] * 299) + (rgb[1] * 587) + (rgb[2] * 114)) / 1000;
-        return this.sanitization.bypassSecurityTrustStyle((yiq >= 128) ? '#f0f0f0' : '#0d0d0d');
-    }
-
-    private hslToRgb(h, s, l) {
-        const c = (1 - Math.abs(2 * l - 1)) * s;
-        const x = c * (1 - Math.abs((h / 60) % 2 - 1));
-        const m = l - c / 2;
-
-        let r, g, b;
-        if (0 <= h && h < 60) {
-            r = c;
-            g = x;
-            b = 0;
-        } else if (60 <= h && h < 120) {
-            r = x;
-            g = c;
-            b = 0;
-        } else if (120 <= h && h < 180) {
-            r = 0;
-            g = c;
-            b = x;
-        } else if (180 <= h && h < 240) {
-            r = 0;
-            g = x;
-            b = c;
-        } else if (240 <= h && h < 300) {
-            r = x;
-            g = 0;
-            b = c;
+    public getForegroundColor(column: TableData, label: string): SafeHtml {
+        const color = column.type.foregroundColors.getColor(label);
+        if (color === '') {
+            return this.sanitization.bypassSecurityTrustStyle('#0d0d0d');
         } else {
-            r = c;
-            g = 0;
-            b = x;
+            return this.sanitization.bypassSecurityTrustStyle(color);
         }
-        return [Math.round((r + m) * 255), Math.round((g + m) * 255), Math.round((b + m) * 255)];
     }
 }